requirements.txt
python-dotenv==0.21.1
colorama==0.4.6
py-cord==1.7.3